fix(tracks): canon short-month date grammar
The /tracks row date column now uses the short-month archival form
("Jul 5, 2026") — VOICE.md's Tabular Rule ("Found Jun 4") and the
"Jun 4, 2026" the rest of the app already ships (formatDateLong). The
partial-precision honesty guard is unchanged. Column tightened to 5.5rem;
the phone-wrap media block is gone (a short-month date is one line at any
width).

+++ apps/web/src/lib/format.ts
@@ -38,12 +38,12 @@ export function formatDateLong(value: string): string {
   return dateLongFormatter.format(new Date(value));
 }
 
-// The full human release date ("July 5, 2026") for the /tracks reference row's date column — the one
-// place a spelled-out month reads, where the row has the horizontal room the tabular feed forms
-// (formatDate/formatDateLong) do not. UTC-pinned like its siblings so SSR matches hydration.
+// The release date ("Jul 5, 2026") for the /tracks reference row's date column — the archival form
+// with the year, matching VOICE.md's Tabular Rule ("Found Jun 4") and the "Jun 4, 2026" the rest of
+// the app already ships (formatDateLong). UTC-pinned like its siblings so SSR matches hydration.
 const releaseDateFormatter = new Intl.DateTimeFormat("en-US", {
   day: "numeric",
-  month: "long",
+  month: "short",
   timeZone: "UTC",
   year: "numeric",
 });
@@ -53,9 +53,9 @@ const releaseDateFormatter = new Intl.DateTimeFormat("en-US", {
 const FULL_DAY = /^\d{4}-\d{2}-\d{2}$/;
 
 /**
- * The /tracks row's date column. A full `YYYY-MM-DD` release date reads as the spelled-out human date
- * ("July 5, 2026"); a partial-precision date stays honest at the YEAR alone (never fabricating a month
- * or day the source never gave — `new Date("2026")` would otherwise render "January 1"); an
+ * The /tracks row's date column. A full `YYYY-MM-DD` release date reads as the archival short-month
+ * date ("Jul 5, 2026"); a partial-precision date stays honest at the YEAR alone (never fabricating a
+ * month or day the source never gave — `new Date("2026")` would otherwise render "Jan 1"); an
  * empty/absent value shows an em dash.
  */
 export function formatReleaseDate(releaseDate: string): string {
